Moved PlayResumeCommandTests to upper class
